+++ src/main/java/com/ibm/cloud/cloudant/kafka/SourceChangesConnector.java
@@ -13,12 +13,11 @@
  */
 package com.ibm.cloud.cloudant.kafka;
 
-import com.ibm.cloud.cloudant.kafka.common.InterfaceConst;
-import com.ibm.cloud.cloudant.kafka.common.utils.JavaCloudantUtil;
-import com.ibm.cloud.cloudant.kafka.connect.CachedClientManager;
-import com.ibm.cloud.cloudant.kafka.connect.SourceChangesConnectorConfig;
-import com.ibm.cloud.cloudant.kafka.connect.ConfigValidator;
-import com.ibm.cloud.cloudant.kafka.connect.SourceChangesTask;
+import com.ibm.cloud.cloudant.kafka.utils.JavaCloudantUtil;
+import com.ibm.cloud.cloudant.kafka.caching.CachedClientManager;
+import com.ibm.cloud.cloudant.kafka.tasks.SourceChangesConnectorConfig;
+import com.ibm.cloud.cloudant.kafka.validators.ConfigValidator;
+import com.ibm.cloud.cloudant.kafka.tasks.SourceChangesTask;
 import org.apache.kafka.common.config.Config;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
@@ -26,7 +25,6 @@ import org.apache.kafka.connect.source.SourceConnector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;